2015-08-08 13:48 UTC+0200 Viktor Szakats (vszakats users.noreply.github.com)   * contrib/hbcrypto/scrypt.c     ! fix extern declarations when built in C++ mode
  * contrib/hbcrypto/3rd/ed25519/ed25519.dif
  * contrib/hbcrypto/3rd/ed25519/fe.c
    ! fix msvc64 warning:
      warning C4146: unary minus operator applied to unsigned type, result still unsigned

  * contrib/hbcrypto/bcrypt.c
    ! fix msvc64 warnings:
      warning C4267: 'argument': conversion from 'size_t' to 'HB_U16', possible loss of data

  * contrib/hbnf/fttext.c
    ! fix msvc64 warning:
      warning C4244: '=': conversion from 'HB_SIZE' to 'int', possible loss of data

  * contrib/hbssl/ssl_inet.c
    ! fix msvc64 warning:
      warning C4244: 'argument': conversion from 'HB_SOCKET' to 'int', possible loss of data

  * contrib/hbtip/mime.c
    ! fix msvc64 warning:
      warning C4244: 'initializing': conversion from 'HB_SIZE' to 'int', possible loss of data
    * use HB_SIZE instead of HB_ISIZ to avoid further warnings
      (I hope this is alright)

  * package/mpkg_win.bat
    ! fix to find 'strip' without relying on mingw being in PATH

  * src/common/hbver.c
    ! fix typo in cast

  * src/rtl/sha1.c
    * silence msvc warning:
      conversion from 'HB_SIZE' to 'sha1_quadbyte', possible loss of data

+++ contrib/hbssl/ssl_inet.c
@@ -317,7 +317,7 @@ static void hb_inetStartSSL( HB_BOOL fServer )
                pStream->blocking = !pStream->blocking;
 
             SSL_set_mode( ssl, HB_SSL_MODE_AUTO_RETRY );
-            iResult = SSL_set_fd( ssl, sd );
+            iResult = ( int ) SSL_set_fd( ssl, sd );
             while( iResult == 1 )
             {
                if( fServer )
